+++ scripts/signpost-fast-travel/actor.lua
@@ -1,7 +1,7 @@
-require("scripts.signpost-fast-travel.checks")
 local AI = require("openmw.interfaces").AI
 local core = require('openmw.core')
 local self = require('openmw.self')
+local common = require("scripts.signpost-fast-travel.common")
 
 local amDead = false
 local combatRegistered = false
@@ -37,6 +37,11 @@ local function onUpdate()
     local curPkg = AI.getActivePackage(self)
     if curPkg and curPkg.target and curPkg.type == "Combat" and not combatRegistered then
         if curPkg.target.recordId ~= "player" then return end
+        -- The Combat AI package can be active on an actor when combat itself
+        -- is not active. Try to account for that with this distance check.
+        if (curPkg.target.position - self.position):length() >= common.COMBAT_END_DISTANCE then
+            return
+        end
         core.sendGlobalEvent('momw_sft_globalRegisterCombat', {entity = self})
         combatRegistered = true
     elseif curPkg and not curPkg.type == "Combat" and combatRegistered then
@@ -45,10 +50,14 @@ local function onUpdate()
 end
 
 local function Died()
-	amDead = true
+    amDead = true
     if combatRegistered then
         deRegisterCombat()
     end
+end
+
+local function distanceEndCombat()
+    combatRegistered = false
 end
 
 return {
@@ -58,5 +67,8 @@ return {
         onSave = onSave,
         onUpdate = onUpdate
     },
-    eventHandlers = {Died = Died}
+    eventHandlers = {
+        Died = Died,
+        momw_sft_distanceEndCombat = distanceEndCombat
+    }
 }